+++ Assets/01_Scripts/Player/Attack/FireCircleDeco.cs
@@ -4,7 +4,7 @@ public class FireCircleDeco : AttackDeco
 {
     public FireCircleDeco(IAttack attack) : base(attack) { }
 
-    public override void FireCircle(int AttackDamage, AttackData data, MonsterPoolManager poolManager, LayerMask layer)
+    public override void FireCircle(float AttackDamage, AttackData data, MonsterPoolManager poolManager, LayerMask layer)
     {
         
         base.FireCircle(AttackDamage, data, poolManager, layer);
